requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.25
httpx==0.27.2
kafka-python==2.0.2
